PR1 and PR2 reviewed changes
