+++ .gitignore
@@ -67,6 +67,7 @@ celerybeat-schedule.*
 # dotenv #
 .env
 .venv
+.rdb
 # virtualenv #
 venv/
 ENV/
