+++ todo_list/settings.py
@@ -142,7 +142,7 @@ USE_I18N = True
 USE_TZ = True
 
 
-LOGIN_URL = 'todo_list:login'
+LOGIN_URL = 'login'
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/5.0/howto/static-files/
