--- todo_list_app/models.py
@@ -1,7 +1,0 @@
-
-# Create your models here.
-
-from .models import Category
-from .models import Task
-from .models import Reminder
-
