+++ todo_list_app/templates/todo_list_app/category_update.html
@@ -4,5 +4,5 @@
   {% csrf_token %}
   <label for="name">Category Name:</label>
   <input type="text" id="name" name="name" value="{{ category.name }}">
-  <button type="submit">Create</button>
+  <button type="submit">Update</button>
 </form>
