+++ todo_list_app/templates/todo_list_app/index.html
@@ -7,10 +7,10 @@
     {% endif %}
 </div>
 {%if user.is_authenticated %}
-  <a href="{% url 'todo_list:logout' %}">logout</a>
+  <a href="{% url 'logout' %}">logout</a>
 {% else %}
-  <a href="{% url 'todo_list:login' %}">login</a>
-  <a href="{% url 'todo_list:register' %}">register</a>
+  <a href="{% url 'login' %}">login</a>
+  <a href="{% url 'register' %}">register</a>
 {% endif %}
 <ul>
   <li><a href="{% url 'todo_list:all_categories' %}">categories</a></li>
